move gc files around
